+++ MaterialWarehousing/src/main/java/com/mit/controller/BoardController.java
@@ -15,12 +15,11 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.mit.domain.Criteria;
 import com.mit.domain.EmailDTO;
-import com.mit.domain.HandleVO;
 import com.mit.domain.PageDTO;
 import com.mit.service.CompanyService;
 import com.mit.service.EmailService;
 import com.mit.service.OrderStatusService;
-import com.mit.service.WareHandlingService;
+import com.mit.service.PartService;
 import com.mit.service.transactionCloseService;
 
 import lombok.AllArgsConstructor;
@@ -33,8 +32,8 @@ import lombok.extern.log4j.Log4j;
 public class BoardController {
 	
 	private transactionCloseService service1;
-	private WareHandlingService whservice;
 	private CompanyService cpservice;
+	private PartService pservice;
 	
 	
 	// 전체 목록 /main(get)	-> /main.jsp
@@ -42,47 +41,6 @@ public class BoardController {
 	public void main() {
 		log.info("main 요청");
 	}
-	
-	//입고처리페이지
-	@GetMapping("wareHandling")
-	public void wareHandling(Model model, Criteria cri) {
-		log.info("wareHandling 요청");
-		if (cri != null) {
-			if (cri.getCompanyName()=="") {
-				cri.setCompanyName(null);
-			}
-			if (cri.getEndDate()=="") {
-				cri.setEndDate(null);
-			}
-			if (cri.getPartName()=="") {
-				cri.setPartName(null);
-			}
-			if (cri.getStartDate()=="") {
-				cri.setStartDate(null);
-			}
-		}
-		model.addAttribute("whList", whservice.getList(cri));
-		model.addAttribute("pageMaker", new PageDTO(cri, whservice.count(cri)));
-	}
-	
-	//입고처리
-	@PostMapping("handleok")
-	public String handleok(HandleVO ho, RedirectAttributes rttr){
-		log.info("입고처리 요청");
-		whservice.handleok(ho);	//입고처리 요청
-		rttr.addFlashAttribute("ware_num", ho.getWare_num());	//입력된 입고번호 전송
-		return "redirect:wareHandling";
-	}
-		
-	//반품처리
-	@PostMapping("handleno")
-	public String handleno(HandleVO ho, RedirectAttributes rttr){
-		log.info("반품처리 요청");
-		whservice.handleno(ho);	//반품처리 요청
-		rttr.addFlashAttribute("ware_num", ho.getWare_num());	//입력된 입고번호 전송
-		return "redirect:wareHandling";
-	}
-
 	
 	// 거래마감  
 	@GetMapping("transactionClose")
@@ -116,8 +74,9 @@ public class BoardController {
 	
 	// 품목조회/등록  -> 조회
 		@GetMapping("productList")
-		public void productList() {
+		public void productList(Model model, Criteria cri) {
 			log.info("productList 요청");
+		model.addAttribute("ProductList", pservice.getList(cri));
 	}
 		
 	// 거래명세서
