+++ MaterialWarehousing/src/main/resources/com/mit/mapper/WareHandlingMapper.xml
@@ -23,6 +23,14 @@
    					AND order_date between TO_TIMESTAMP(#{startDate}, 'YYYY-MM-DD')-1 and TO_TIMESTAMP(#{endDate}, 'YYYY-MM-DD')+1 
    					<!-- 날짜 검색 기능 개선 (between은 양쪽 날짜가 포함되지 않아 -1, +1 함)-->
    				</if>
+   				<if test='recentDate !=null and recentDate.equals("week")'>
+   					<![CDATA[ AND order_date >= sysdate-7  ]]>
+   					<!-- 최근 일주일  -->
+   				</if>
+   				<if test='recentDate !=null and recentDate.equals("month")'>
+   					<![CDATA[ AND order_date >= sysdate-30 ]]>
+   					<!-- 최근 한달  -->
+   				</if>
    			</trim>
     	</trim>
     </sql>
